+++ vm/stdlib/compiled/release_notes.md
@@ -1,5 +1,11 @@
 # Starcoin Stdlib Release Notes
 
 Version 1
+
     - release for Starcoin 1.0
+
+Version 2
+
+    - for test
+    - init_script: "cargo run -- -v 2 -m PackageTxnManager -f convert_TwoPhaseUpgrade_to_TwoPhaseUpgradeV2 -a 0x1"